% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{labels.countries}
\alias{labels.countries}
\title{Compute a unique \code{sf} to optimize the position of labels}
\usage{
\method{labels}{countries}(object, ..., var)
}
\arguments{
\item{object}{a \code{countries} object`,}

\item{\dots}{further arguments (currently unused),}

\item{var}{a character vector indicating the entities that should
be labeleld, it can be \code{country}, \code{capital} and \code{towns}}
}
\value{
a \code{sf} containing:
\itemize{
\item \code{name} the names of the entities,
\item \code{type} the type of the entity (either \code{country}, capital\code{and}towns`)
\item \code{point} the coordinate of the points (obtained using
\code{st_point_on_surface} for countries)
}
}
\description{
Displaying labels on a map is complicated because of serious
overlapping problems. Labels for different kinds of entities can be
put in a unique \code{sf}
}
