% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ne_countries}
\alias{ne_countries}
\title{Countries of naturalearth}
\format{
a \code{sf} containing
\itemize{
\item \code{iso2}: two letters identifier of the country,
\item \code{iso3}: two letters identifier of the country,
\item \code{type}: either \code{"main"} (the main part of a sovereign country, the
whole country for most of them), \code{"part"}, \code{"dependency"} or \code{"indeterminate"}
\item \code{country}: the name of the entity,
\item \code{sovereign}: the sovereign country the entity belongs to,
\item \code{capital}: the name of the capital of the country (NA for parts
and dependencies)
\item \code{status}: United Nations' status
\item \code{en}, \code{fr}, \code{de}, \code{es}, \code{it}: the name of the country in
different languages
\item \code{region}: the name of the region (United Nations' definition)
\item \code{subregion}: the name of the subregion (United Nations'
definition)
\item \code{wbregion}: the name of the region (World Bank's definition)
\item \code{pop}: the population of the entity,
\item \code{gdp}: currently undocumented
\item \code{economy}: economic group
\item \code{income}: income groupe
\item \code{polygon} a geometry column containing the administrative borders
\item \code{point} a geometry column containing the point coordinate of the
capital
}
}
\description{
A \code{sf} containing 299 countries (in a large sense), either the main
parts of sovereign countries, parts or dependencies
}
\keyword{datasets}
