% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ne_towns}
\alias{ne_towns}
\title{Populated places of naturalearth}
\format{
a \code{sf} containing
\itemize{
\item \code{id}: the id of the  country,
\item \code{name}: the name of the city,
\item \code{capital}: a boolean, \code{TRUE} for a capital
\item \code{pop}: the population of the city,
\item \code{point}: a point \code{sfc} containing the coordinates of the city
}
}
\description{
A \code{sf} containing 7342 cities
}
\keyword{datasets}
