% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{left_join}
\alias{select}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:mutate-joins]{left_join}}, \code{\link[dplyr]{select}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}
}}

