% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{slave_trade}
\alias{slave_trade}
\title{Slave trade and economic development}
\format{
a tibble containing:
\itemize{
\item country: the country name
\item region: one of 'north', 'east', 'central', 'south', 'west'
\item disteq: distance from equator
\item longitude: longitude
\item area: area in thousands of km squared
\item pop: average population during the slave trade period
\item coastline: log coastlines divided by the country area
\item island: island indicator
\item islam: percent islamic
\item colony: previous colonizator, one of 'none', 'uk', 'france', 'portugal', 'belgium', 'spain', 'germany', 'italy'
\item legor: legal origin, one of 'french' and 'british'
\item gdp: log real gdp per capita in 2000
\item slaves: number of slaves
\item slavesarea: number of slaves divided by the country area
\item humidmax: average maximum humidity
\item rainmin: lowest month rainfall
\item lowtemp: average minimum temperature
\item gold: log gold production per inhabitant
\item oil: log oil production per inhabitant
\item diamond: log diamonds production per inhabitant
\item atlantic: distance to the atlantic ocean
\item indian: distance to the indian ocean
\item redsea: distance to the red sea
\item sahara: distance to sahara
}
}
\source{
Nathan Nunn's website \url{https://nathannunn.arts.ubc.ca/}
}
\description{
a cross-section of 52 countries from 2000
}
\references{
\insertRef{NUNN:08}{necountries}
}
\keyword{dataset}
