% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sp_solow}
\alias{sp_solow}
\title{Solow's growth model with spatial correlation}
\format{
a tibble containing:
\itemize{
\item name: the name of the country
\item code: the id of the country
\item gdp60: per capita gdp in 1960
\item gdp95: per capita gdp in 1995
\item saving: saving rate
\item labgwth: growth rate of the labor force
}
}
\source{
JAE data archive
}
\description{
a cross-section of 91 countries from 1995
}
\references{
\insertRef{ERTU:KOCH:07}{necountries}
}
\keyword{dataset}
