% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{utm}
\alias{utm}
\title{Universal Transverse Mercator projection}
\usage{
utm(x)
}
\arguments{
\item{x}{either an integer (from 1L to 60L) or a `sf}
}
\value{
a character string

a character (a crs i, the 'proj4string' format)
}
\description{
\code{utm} returns the relevant UTM crs (in the 'proj4string' form)
}
\details{
There is one utm projections for each of the 60 zones that divide
the world. The zone can be indicated as an integer (ie 12L, and not
12) or can be computed from a \code{sf} object
}
\examples{
we <- countries("Western Europe")
utm(we)
utm(32L)
}
