% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.countries}
\alias{plot.countries}
\title{Basic plot function for \code{countries} objects}
\usage{
\method{plot}{countries}(
  x,
  ...,
  labels = NULL,
  fill = NULL,
  capital = NULL,
  centroid = NULL,
  towns = NULL,
  bks = NULL,
  n = 6,
  size = c(1, 10),
  shape = c(15, 16),
  style = NULL,
  palette = NULL,
  bw = FALSE,
  oceans = "lightblue",
  background = "lightgrey"
)
}
\arguments{
\item{x}{a \code{countries} object,}

\item{\dots}{further arguments (currently unused)}

\item{labels}{a character vector containing the variables that
should be labeled: \code{country}, \code{capital} and/or \code{towns}}

\item{fill}{a variable use to fill countries' polygons}

\item{capital, centroid, towns}{a variable associated with the shape
or the size of points; for towns, the only available variable
is "pop", it town is a boolean fixed points are used with a
specific shape for capitals if shape is of length 2}

\item{bks}{an optional vector of breaks in order to use a
continuous variable for fill}

\item{n}{the number of class (passed to \code{classIntervals})}

\item{size}{either an integer for a fixed size or a couple of
integer to define the range of the sizes of the points}

\item{shape}{either one integer for a fixed shape or a couple of
integer to get a specific shape for capitals}

\item{style}{the style (passed to \code{classIntervals})}

\item{palette}{the palette (selected in \code{scale_fill_brewer})}

\item{bw}{a boolean, if \code{TRUE}, a black and white map is produced}

\item{oceans, background}{a character indicating the color for the
oceans and the background}
}
\value{
a \code{gg} object.
}
\description{
As the \code{plot} method of \code{sf}, this function is intended to obtain
quickly a map for a set of countries. Countries' boundaries are
represented and filling can be used, some cities can also be
represented and labels can be added. \code{ggplot} is used and should be
used directly when more enhanced maps are required
}
\examples{
we <- countries("Western Europe")
plot(we)
}
